Use Criterion for benchmark purposes because bench will be deprecated
