+++ Cargo.toml
@@ -20,6 +20,7 @@ snafu = "0.5.0"
 structopt = "0.3"
 
 [dev-dependencies]
+criterion = "0.3.5"
 pretty_assertions = "0.6.1"
 
 [lib]
@@ -30,3 +31,7 @@ path = "src/lib.rs"
 name = "imagecli"
 path = "src/main.rs"
 doc = false
+
+[[bench]]
+name = "benchmark"
+harness = false
